fix : main page rebuild
