+++ src/entities/lab/ui/lab-link.tsx
