+++ src/entities/lab/ui/lab-member.tsx
