+++ src/widgets/lab/index.ts
@@ -1,11 +1,6 @@
-export { LabLink } from './ui/lab-link'
-export { LabMember } from './ui/lab-member'
-
-export { SideNotice } from './ui/side/side-notice'
-export { SideData } from './ui/side/side-data'
-export { SideManagement } from './ui/side/side-management'
-export { SideRequest } from './ui/side/side-request'
-export { SideRental } from './ui/side/side-rental'
-export { SideMember } from './ui/side/side-member'
-export { SideWork } from './ui/side/side-work'
-export { SideWeek } from './ui/side/side-week'
+export { LabMenuWrapper } from './ui/menu/lab-menu-wrapper'
+export { LabLabInfoSection } from './ui/pc/section/lab-lab-info-section'
+export { LabUserDataSection } from './ui/pc/section/lab-user-data-section'
+export { LabWorkingSection } from './ui/pc/section/lab-working-section'
+export { LabMobileTopSection } from './ui/mobile/section/lab-mobile-top-section'
+export { LabMobileCenterSection } from './ui/mobile/section/lab-mobile-center-section'
