+++ src/widgets/lab/ui/mobile/section/lab-mobile-center-section.tsx
@@ -1,0 +1,16 @@
+'use client'
+
+import { useState } from 'react'
+import { LabLink } from '../../../../../entities/lab/ui/lab-link'
+import { LabLinkModal } from '../../modal/lab-link-modal'
+
+export const LabMobileCenterSection = () => {
+  const [isOpen, setIsOpen] = useState<boolean>(false)
+
+  return (
+    <>
+      <LabLink setIsOpen={setIsOpen} />
+      <LabLinkModal isOpen={isOpen} onClose={() => setIsOpen(false)} />
+    </>
+  )
+}
