+++ src/widgets/lab/ui/mobile/section/lab-mobile-top-section.tsx
@@ -1,0 +1,13 @@
+'use client'
+
+import { LabInfo } from '@/entities/lab'
+import { UserInfo } from '@/entities/user'
+
+export const LabMobileTopSection = () => {
+  return (
+    <>
+      <UserInfo isMain isShadow />
+      <LabInfo className={'p-5 md:p-6'} isShadow />
+    </>
+  )
+}
