+++ src/widgets/lab/ui/modal/lab-link-modal.tsx
