+++ src/widgets/layout/menu/index.ts
@@ -1,3 +1,2 @@
 export { PageGnb } from './ui/page-gnb'
 export { MobileBottomBar } from './ui/mobile-bottom-bar'
-export { LabMenuWrapper } from '../menu/ui/lab-menu-wrapper'
