+++ src/widgets/layout/modal/index.ts
@@ -1,5 +1,4 @@
 export { StuLabModal } from './ui/stu-lab-modal'
-export { LabLinkModal } from './ui/lab-link-modal'
 export { Dialog } from './ui/dialog'
 export { Alert } from './ui/alert'
 export { Confirm } from './ui/confirm'
